digital-images/raster-vs-vector/requirements.txt
numpy==1.26.0
matplotlib==3.8.0
svgwrite==1.4.3
Pillow==9.5.0
